Change aspect ratio from double to float.
The field on the image struct was already a float.

+++ src/dtgtk/thumbnail.h
@@ -130,7 +130,7 @@ typedef struct
 
   // specific for culling and preview
   gboolean zoomable;   // can we zoom in/out the thumbnail (used for culling/preview)
-  double aspect_ratio; // aspect ratio of the image
+  float aspect_ratio;  // aspect ratio of the image
 
   // difference between the global zoom values and the value to apply to this specific thumbnail
   float zoom;     // zoom value. 1.0 is "image to fit" (the initial value)
